LayoutTests/platform/win/editing/selection/5825350-2-expected.txt

CONSOLE MESSAGE: line 14: ReferenceError: Can't find variable: textInputController
This tests for a bug where moving the caret right towards a non-editable pocket of an editable region would make the caret disappear. The caret should be just after the 'y' in "Sally".
Bob ; Sally
